requirements.txt
# requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic[email]
pydantic-settings
passlib
python-jose[cryptography]
python-multipart
opencv-python
facenet_pytorch
torch
numpy
scikit-learn
pandas
surprise
joblib
psycopg2-binary